Access to windows timer internal representation added.
git-svn-id: http://b035a0a.online-server.cloud/svn/Acf@563 5144ca80-2444-0410-90ef-8152f95a0219

+++ Include/iwin/CTimer.cpp
@@ -40,12 +40,12 @@ void CTimer::Start()
 
 double CTimer::GetElapsed() const
 {
-	I_ASSERT(sizeof(long long) == sizeof(LARGE_INTEGER));
+	I_ASSERT(sizeof(unsigned long long) == sizeof(LARGE_INTEGER));
 
-	long long endCounter;
+	unsigned long long endCounter;
 	::QueryPerformanceCounter((LARGE_INTEGER*)&endCounter);
 
-	return (endCounter - m_startCounter) / double(s_timerFrequence);
+	return (long long)(endCounter - m_startCounter) / double(s_timerFrequence);
 }
 
 
@@ -79,7 +79,7 @@ double CTimer::GetTimerResolution() const
 
 // static attributes
 
-long long CTimer::s_timerFrequence;
+unsigned long long CTimer::s_timerFrequence;
 bool CTimer::s_isTimerFrequenceValid = (::QueryPerformanceFrequency((LARGE_INTEGER*)&CTimer::s_timerFrequence) != 0);
